requirements.txt
openai>=1.0.0
pinecone-client>=3.0.0
langchain>=0.1.0
streamlit>=1.25.0
python-dotenv>=1.0.0
numpy>=1.23.0
